requirements.txt
python-telegram-bot==20.7
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
pydub==0.25.1
SpeechRecognition==3.10.0
pyaudio==0.2.11

